requirements.txt
google-generativeai
PyMuPDF
easyocr
pdf2image